Improve game flow, ML integration and auth
Add /logout endpoint and invalidate refresh token on logout
Update docker-compose.yml to build the ML service image
Implement game termination on the 18th turn and set endTime
Adjust dice multiplier coefficients to a smoother 0.7–1.3 range
Add evaluatePresentation() endpoint (presentation evaluation without dice)
Add score column to Game entity and database migration
Refactor stage update logic to bump stage every 6 turns

+++ backend/app/src/main/java/com/startupgame/dto/game/ResourceDelta.java
@@ -9,5 +9,5 @@ public class ResourceDelta {
     private int motivation;
     private int technicalReadiness;
     private int productReadiness;
-    private double money;
+    private long money;
 }
